Changes: forget configuration of uninstalled applications
RevBy: Pauli Lehtinen
Details: when an application is removed from the system, the corresponding
launcher button in the quick launcher bar was already removed. But the
configuration was remembered, so if the same application was reinstalled,
the launcher button would reappear. This was not a desired behaviour, so
this patch will change the behaviour so that the configuration is forgotten
and the launcher button will not reappear.