Changes: package building working again
RevBy: TrustMe
Details:
Return the english translation for the weather applet - there's a debian package for it anyway.

Don't include the engineering english .ts files to the engineering english packages (although
this is recommended by the localization packaging guidelines). For weatherapplet this probably
makes no difference but duihome is production code so there we should follow the guidelines.
However, there are no localizable strings currently in duihome, so it shouldn't matter right now.