Changes: SwitcherButtonView listens to visibility notify signals directly
RevBy: Pauli Lehtinen
Details: SwitcherButtonView now has its own XEventListener which is tailored
for SwitcherButtonView's needs. The listener notices VisibilityNotify X events
and causes the button to be updated if needed.

SwitcherButtonView was the only client for the windowVisibilityChanged signal
in Switcher so that got now removed since nobody needs that anymore.