New: layout visualization wrapper
RevBy: Artem Egorkine
Details: this helper class wraps the visualization of a linear layout.
It transforms the first and last item in the layout to the other end of
layout when told to do so. This can be used to have an effect of making
the widget with the layout to appear continuous.