Fixes: NB#206700 - minimizing window, meegotouchhome crashes
RevBy: Artem Egorkine
Details: QPixmap::drawPixmap might throw a std::bad_alloc if the X pixmap
that is attempted to be drawn is already released. This happens when the
backing store of the application thumbnail has changed but SwitcherButtonView
hasn't yet received that information or hasn't yet adapted to the change.
In this case a draw of the background is skipped but it's only temporary,
the view will start to use the new pixmap in a short moment (probably when
the event loop gets to the point that it informs the view about the change).