Fixes: NB#208911 - Edit mode of Recents (Switcher) view should exit when status menu or any dialog comes on top.
RevBy: Eetu Lehmuvuo
Details: Changed access specifier for windowMonitor member of Switcher from private to protected so it can be used in derived classes. Added signal anyWindowOnTopOfOwnWindow(WindowInfo) to Switcher which basically does the same thing as the fullscreenWindowOnTopOfOwnWindow() but the signal is emitted for all window types and WindowInfo is passed as parameter. Added test cases for Switcher anyWindowOnTopOfOwnWindow(WindowInfo) signal