try fix damage event processing.
under certain circumstances, we must invalidate and refetch the window
pixmap, so listen to X events for these circumstances.

what isn't addressed in this patch (but should probably be done) is to
ignore damage events whilst the desktop isn't in view, and attempt to
catch up when it is restored.