Support for remote actions
The desktop notification specification only specifies a D-Bus signal which is sent when a notification action is invoked. This requires the target process to be running in order to catch the D-Bus signal. In a mobile device the target application is not necessarily running. This patch implements a Nemo hint which allows D-Bus calls to be associated with actions like in Harmattan. This way the target application can be started on demand. The notification removability is not related to whether the notification has actions defined or not. Also, remote actions can be defined without defining visual representations for the actions.