Battery status indicator backend
The StatusIndicator class provides a generic mechanism to implement status indicators
on the home screen. The indicator can display a single value from a list of values or
animate through a range of values from a list of values. The indicator may set
its mode which allows the view component to define different behaviors for different modes.
The BatteryStatusIndicator class uses these facilities to provide data for a battery
status indicator.