[homescreen] Fixes JB#2844, JB#2845: replace GridView in switcher with a better layout
We use a Grid with custom-set margins, positions, etc, for the best
possible look. This isn't scalable at all, but it's better than nothing.

Pixmaps aren't ideally scaled due to the design, the agreement is that
we need to look into cropping the pixmaps. This will need changes in
lipstick, and is thus filed as JB#2989