Simplify pixmap refresh
Instead of trying to figure out the magical incantation to maintain
damage and an updated pixmap, just give in and refresh it when
homescreen-on-top status changes.

This fixes windows occasionally being shown with incorrect content when
a particular blue moon is visible on Cthulu's birthday (17th of November
according to the calendar of the orthodox serbian church.)

This also has another good side effect: createDamage would occasionally
be called without destroying the old damage first, so this, in theory,
fixes a leak.