[lipstick] add connection selector
Connection selector allows the user to choose a connection when none
is available.
It needs an appropriate qml file, and a patch to connman's dnsproxy that
sends a signal when there is no connection available, and one
is requested.

[lipstick] add connection selector

[lipstick] change email address to something else

fix spelling error

[lipstick] remove other mention of showButtons.

add window resize, otherwise this does not show up

review changes/fixes