[lipstick] Cap damage event processing when the homescreen is backgrounded.
It makes no sense to constantly process damage events when the homescreen is not
the center of focus, this just drains resources. This is especially noticable if
the client is not using the MeeGo graphicssystem, as pixmap updates are
considerably more expensive (which can drag client performance down to 40fps or
worse).