[lipstick] Queue a window update on map/unmap/surface damage
Calling QQuickItem::update is not always enough to guarantee
QQuickWindow frame swap, which is in turn is used to trigger
frame callbacks to clients. If a client expects a callback
but due to, for example, scene graph sync/paint optimizations,
a frame swap is skipped, it might cause a deadlock in the client
until the compositor draws a frame. Thus, schedule a repaint
always.