[lipstick] Send a cancel event when the window is "lost" during touch
If a LipstickCompositorWindow is hidden, disabled, or touch is disabled
while a touch series is happening, future touch events will not be
received by the surface and the touch point pool may never be cleared.
For this reason, generate a TouchCancel event if the window has touch
focus and gets hidden, disabled, or touch events are disabled.