[launchermonitor] Rewrite monitoring of .desktop and icon files
Fixes problems where the launcher wouldn't find the icons in the right
location for some cases. Right now, this hard-codes the icon size to
86x86, but it can easily be extended later to pick a given icon size
and/or monitor multiple directories.

Monitoring of multiple directories with .desktop files is currently not
supported, but can easily be added (it wasn't utilized before as well).