[rpm] Create icon directory, so it can always be monitored
We need to create the icon directory in the RPM package, so that it will
always exist when lipstick is started, otherwise monitoring of the icon
directory will fail when no third party apps are installed.

Thanks to Vesuri and Reto for debugging the issue, testing the fix and
coming up with the solution.