[lipstick] Fix for Device lock is sometimes activated when using device
Recent chages in mce stopped sending qmLock signals while suspended, which
breaks the logic in devicelock handling. In some cases like on android
games and videoplayer triggered qmActivity inactive signal which
triggered devicelock timer to start. And then display ON signal came the
timer was not resetted because qmActivity and qmLocks mce signaling was
changed. Which triggered devicelock code to be shown when swiping away
from long usage applications.

With this change we add handling to check possible lockTimer reset on
display ON signal. And we no longer query for qmLocks state when
lockingDelay == 0.