[launchermodel] Iterate over list of launchers in reverse order
Reduces best case time of lookups for new installations (icon is
appended at the end of the list) from O(n) to O(1). If looking up an
item in the list will become a bottleneck, we can do a hash over the
whole list with keys being the package name.