[path] Lipstick QmlPath API for installing files to the filesystem
Remove the assumption that QML files are always put into Qt Resources.

This allows implementations of homescreens to install files in the
filesystem and reference those files directly, as well as override the
path to the QML files during development easily.

Homescreens should use:

 - QmlPath::append() / QmlPath::prepend() to tell Lipstick about the
   paths to the QML files (absolute file paths or starting with ":"
   for paths to Qt Resources)
 - QmlPath::to() with the basename of the QML file to look up the QUrl
   of a QML file; this will be resolved using the existing paths

There is a fallback implemented that will add ":/qml" the first time
QmlPath::to() is called if no QmlPath::append() has been called before.
This allows legacy homescreens to function without change, although it
is recommended that homescreens start using QmlPath::append() and
QmlPath::to() for better integration.

Lipstick internally uses QmlPath::to() everywhere it looks up QML files.

The header "lipstickqmlpath.h" is installed as part of the public API;
version bumped to 0.23.0, which homescreen should depend on when using
the new QmlPath API and the header file.