[lipstick] Resolve desktop file for link type exec
Exec of link should not be used for launching but
can be used as an identity.

Note: When using exec as an identity and there is one desktop file for the
"normal" application launching and another one that has registered the
uri scheme, the value might not be suitable for identity usage.

This should be pushed even deeper, but don't know yet where.