[lipstick[ Implement basic hardware compositor support in lipstick.
On the rendering thread in lipstick, we install a custom scene graph
rendering stage. In this stage, we look through the scene graph for
hardware compatible nodes and put these into their own block.

We're currently only supporting in-process images (for background
images) without rotation (for now) and no scaling. The custom render
stage will look at the transform of these images and decide if it can
be sent to the hwc or not.

We can talk to the hwc using an abstraction API called HwcInterface
which we can query from the QPA plugin. The API is located in
'hwcinterface.h' and is a direct copy of the same file inside
the QPA plugin. It has been done like this to have some API
while keeping this out of the public API. There is a version
name string to prevent mismatches between the plugin and the
hwc.

Once we've identified buffers to draw using the hwc we create a
HwcInterface::LayerList which contains the buffer handles. The layer
list can also include GL rendering which will stack on top. We
schedule this to the hwc and will be notified some time later if it is
accepted or not. Until it is, we render all GL as normal. Once
accepted, we switch to swapping the layer list instead of using
eglSwapBuffers().

The feature is enabled by setting LIPSTICK_HARDWARE_COMPOSITOR=1
in the environment.