[lipstick] Make buffer releasing API slightly more flexible
Deleting a QObject works fine for the HwcImage where the resource is
a QSGTexture, but to support window surface releasing, we would want
it to be a bit more flexible. Use a callback + void* data instead.

This also fixes a bug where opacity changes would fail to toggle
HWC on or off.