[lipstick] Enable use of HWC for window surfaces.
The feature is disabled by default and needs
LIPSTICK_HARDWARE_COMPOSITOR_WINDOWSURFACES=1 in the environment to be
enabled. It also requires a libhybris EGL extension and that client
buffers are created with the HWCOMPOSER usage bit set to work
properly.

It works much the same way as the HwcImage, in that we populate the
scene graph with a HwcNode which in turn contains the wayland client's
texture in a subnode. The render stage can figure out if the buffer
can go through HWC or needs GL composition.

Releasing is a two-stage operation, where we first destroy the texture
on the render thread during the sync phase when a buffer has been
replaced with a new one. At the same time, we tell the HwcRenderStage
to notify us when the buffer is released by the HWC and follow that up
by posting an event to the GUI thread to handle the wayland buffer
release there.