[lipstick] Ensure correct sorting of notifications. Contributes to MER#991
The sorting function must provide a strict weak ordering depending
on the properties of the sorted notifications independent of their
current position in the list.  Use replacesId as the discriminator
when timestamp is equivalent.

Also, do not compare a notification to itself in the list, as the
properties may have been updated since it was sorted into that
position.