[lipstick] Disable mouse event synthesis. Fixes JB#24350
When a touch event is ignored, it causes QtQuick to try a synthesized
mouse event. Disable event synthesis (checks internally for
Qt::LeftButton) to prevent entering zone of undefined behavior, as
Qt will keep on sending further touch events to the item, and we are
not ready to handle them.

This could be an issue in Qt -- touch events probably should not be
delivered to an item that accepted a synthesized mouse press.