[lipstick] Send the frame callbacks when the display is off. Fixes MER#1004
When the display is off we don't repaint so we don't normally send frame
callbacks, which means a client calling eglSwapBuffers() will wait until
the screen comes back on. To prevent that we send the frame callbacks when
a client commits a surface, after a 100ms delay to prevent the clients going
full speed when they should not be drawing.