[lipstick] Re-trigger feedback on notification update. Contributes to MER#1133
When a notification containing a feedback directive is updated, the
feedback should be restarted.

The feedback player used to rely on the old semantics to prevent
feedback from being triggered during lipstick restart.  Instead, use
a hint of type 'x-nemo-restored' to indicate that a notification
has been restored and not intentionally modified, and suppress both
feedback and preview when a notification is in this state.